+++ docs/changelog.rst
@@ -25,6 +25,7 @@ Under development
 * Improvement: Added builder :ref:`needumls_builder` and config option :ref:`needs_build_needumls` in `conf.py`.
 * Improvement: Added `filter` function for :ref:`needuml`.
 * Improvement: Renamed jinja function `need` to `flow` for :ref:`needuml`.
+* Improvement: Added directive :ref:`needarch`.
 
 1.0.1
 -----